Only download the executable corresponding to your operating system

+++ get_results_data.sh
@@ -2,8 +2,9 @@
 mkdir -p results
 if [ ! -d "results/user_study_dataset" ]; then
 	cd results
-	echo "Downloading result files"
+	echo "Downloading result files."
 	wget https://zenodo.org/record/321599/files/user_study_dataset.zip
+	echo "Extracting result files, please wait"
 	unzip -q user_study_dataset.zip
 	rm user_study_dataset.zip
 	cd ..
@@ -11,16 +12,14 @@ fi
 
 
 mkdir -p bin
-if [ ! -d "bin/ReplayInterfaceLinux" ]; then
+if [ "$(expr substr $(uname -s) 1 5)" == "Linux" ] && [ ! -d "bin/ReplayInterfaceLinux" ]; then
 	cd bin
 	echo "Downloading graphical interface binary for Linux system"
 	wget https://github.com/3rdHand-project/PosturalOptimizationDataReader/releases/download/v1.0/ReplayInterfaceLinux.tar.gz
 	tar -xzf ReplayInterfaceLinux.tar.gz
 	rm ReplayInterfaceLinux.tar.gz
 	cd ..
-fi
-
-if [ ! -d "bin/ReplayInterfaceMac" ]; then
+elif [ "$(uname)" == "Darwin" ] && [ ! -d "bin/ReplayInterfaceMac" ]; then
 	cd bin
 	echo "Downloading graphical interface binary for Mac system"
 	wget https://github.com/3rdHand-project/PosturalOptimizationDataReader/releases/download/v1.0/ReplayInterfaceMac.tar.gz
